services/api/pyproject.toml
[project]
name = "discipline-api"
version = "0.0.0"
description = "Discipline OS — backend modular monolith"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Discipline OS" }]

dependencies = [
    "fastapi >=0.115.0,<0.120",
    "uvicorn[standard] >=0.32.0,<0.34",
    "gunicorn >=23.0.0",
    "pydantic >=2.9.0,<3.0",
    "pydantic-settings >=2.5.0",
    "asyncpg >=0.30.0",
    "sqlalchemy >=2.0.35",
    "alembic >=1.13.3",
    "redis >=5.1.0",
    "rq >=2.0.0",
    "httpx >=0.27.0",
    "python-jose[cryptography] >=3.3.0",
    "cryptography >=43.0.0",
    "boto3 >=1.35.0",
    "opentelemetry-api >=1.27.0",
    "opentelemetry-sdk >=1.27.0",
    "opentelemetry-exporter-otlp-proto-http >=1.27.0",
    "opentelemetry-instrumentation-fastapi >=0.48b0",
    "structlog >=24.4.0",
    "anthropic >=0.39.0",
    "stripe >=11.0.0",
    "pgvector >=0.3.6",
]

[project.optional-dependencies]
dev = [
    "pytest >=8.3.0",
    "pytest-asyncio >=0.24.0",
    "pytest-cov >=5.0.0",
    "ruff >=0.6.9",
    "mypy >=1.11.0",
    "types-redis",
    "hypothesis >=6.112.0",
    "pact-python >=2.2.0",
    "locust >=2.31.0",
    "bandit >=1.7.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/discipline"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "I", "N", "W", "UP", "B", "A", "C4", "PT", "SIM",
    "TCH", "RUF", "S",
]
ignore = ["E501", "RUF002", "RUF003", "TC001", "TC002", "TC003"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "S106", "PT011", "PT018", "PT006", "B017", "RUF005", "N802", "PT012", "PT017", "C408", "SIM108", "F841", "N812"]

[tool.mypy]
python_version = "3.12"
strict = true
packages = ["discipline"]
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers --strict-config"

[tool.coverage.run]
branch = true
source = ["discipline"]
